refactor phase diagram file structure by consolidating energies.ts+convex-hull.ts into thermodynamics.ts and ternary.ts+quaternary.ts into barycentric-coords.ts
- Removed `arity.ts`, `convex-hull.ts`, and `energies.ts` files, consolidating their functionality into `types.ts` and `thermodynamics.ts`.
- Added `barycentric-coords.ts` for handling barycentric coordinate transformations and calculations for 3D and 4D phase diagrams.
- Renamed helper functions in `helpers.ts` for clarity and consistency.
- Adjusted imports across components and rename test files to match
