+++ src/lib/phase-diagram/index.ts
@@ -1,10 +1,8 @@
-export * from './convex-hull'
-export * from './energies'
+export * from './barycentric-coords'
 export { default as PhaseDiagram2D } from './PhaseDiagram2D.svelte'
 export { default as PhaseDiagram3D } from './PhaseDiagram3D.svelte'
 export { default as PhaseDiagram4D } from './PhaseDiagram4D.svelte'
 export { default as PhaseDiagramControls } from './PhaseDiagramControls.svelte'
 export { default as PhaseDiagramInfoPane } from './PhaseDiagramInfoPane.svelte'
-export * from './quaternary'
-export * from './ternary'
+export * from './thermodynamics.ts'
 export * from './types'
